Buy dollar logic
Create functions that:
Calculate the total amount in pesos to be paid, with the possibility of calculating with dollars or pesos.
Calculate and split the taxes to pay.
Check that the final amount not to exceed the money account.
Setting the date on foreign currency purchases in progress...

Change random values on generateMoney().